QXmppClient: Reset Stream Management package cache on JID changes
The stream management mechanisms cache sent packages until the client
receives an acknowledgement from the server. When the connection gets
lost, the client resends all packages from the last connection that
have not been acknowledged.

And here comes the problem: When connecting with a different JID, the
client still resends all packages from the last connection. Packages
that were never intended to be sent from another account / to another
server.

This commit fixes this behaviour by resetting the package cache, when
the JID changes.